music-recommendation/requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
scikit-learn
python-dotenv
